Move Constraints codegen service into constraints package
